Extract `test_transform`
Part of #22
