+++ robotools/tests.py
@@ -7,7 +7,7 @@ import numpy
 import pytest
 
 import robotools
-from robotools import evotools, liquidhandling, transform
+from robotools import evotools, liquidhandling
 
 
 class TestStandardLabware(unittest.TestCase):
@@ -3102,150 +3102,6 @@ class TestFunctions(unittest.TestCase):
             "destination",
             evotools._optimize_partition_by(ST, DT, "destination", "Trough source and destination"),
         )
-        return
-
-
-class TestWellShifter(unittest.TestCase):
-    def test_identity_transform(self) -> None:
-        A = (6, 8)
-        B = (8, 12)
-        shifter = transform.WellShifter(A, B, shifted_A01="A01")
-
-        original = ["A01", "C03", "D06", "F08"]
-        expected = ["A01", "C03", "D06", "F08"]
-        shifted = shifter.shift(original)
-        numpy.testing.assert_array_equal(expected, shifter.shift(original))
-        numpy.testing.assert_array_equal(shifter.unshift(shifted), original)
-        return
-
-    def test_center_shift(self) -> None:
-        A = (6, 8)
-        B = (8, 12)
-        shifter = transform.WellShifter(A, B, shifted_A01="B03")
-
-        original = ["A01", "C03", "D06", "F08"]
-        expected = ["B03", "D05", "E08", "G10"]
-        shifted = shifter.shift(original)
-        numpy.testing.assert_array_equal(expected, shifter.shift(original))
-        numpy.testing.assert_array_equal(shifter.unshift(shifted), original)
-        return
-
-    def test_boundcheck(self) -> None:
-        A = (6, 8)
-        B = (8, 12)
-
-        with self.assertRaises(ValueError):
-            transform.WellShifter(A, B, shifted_A01="E03")
-
-        with self.assertRaises(ValueError):
-            transform.WellShifter(A, B, shifted_A01="B06")
-        return
-
-
-class TestWellRotator(unittest.TestCase):
-    def test_init(self) -> None:
-        rotator = transform.WellRotator(original_shape=(7, 3))
-        self.assertEqual(rotator.original_shape, (7, 3))
-        self.assertEqual(rotator.rotated_shape, (3, 7))
-        return
-
-    def test_clockwise(self) -> None:
-        A = (6, 8)
-        rotator = transform.WellRotator(A)
-
-        original = ["A01", "C03", "D06", "F08", "B04"]
-        expected = ["A06", "C04", "F03", "H01", "D05"]
-        rotated = rotator.rotate_cw(original)
-        numpy.testing.assert_array_equal(expected, rotated)
-        return
-
-    def test_counterclockwise(self) -> None:
-        A = (6, 8)
-        rotator = transform.WellRotator(A)
-
-        original = ["A01", "C03", "D06", "F08", "B04"]
-        expected = ["H01", "F03", "C04", "A06", "E02"]
-        rotated = rotator.rotate_ccw(original)
-        numpy.testing.assert_array_equal(expected, rotated)
-        return
-
-
-class TestWellRandomizer(unittest.TestCase):
-    def test_init(self):
-        randomizer = transform.WellRandomizer(original_shape=(1, 4), random_seed=13)
-        assert randomizer.original_shape == (1, 4)
-        assert randomizer.random_seed == 13
-        numpy.testing.assert_array_equal(randomizer.randomized_wells, ["A02", "A04", "A01", "A03"])
-        return
-
-    def test_randomize_wells(self) -> None:
-        A = (6, 8)
-        S = 13
-        randomizer = transform.WellRandomizer(A, S)
-        original = ["A01", "A02", "A03", "A04", "A05", "A06"]
-        expected = ["A01", "F02", "A05", "A07", "B07", "D06"]
-        randomized = randomizer.randomize_wells(original)
-        numpy.testing.assert_array_equal(expected, randomized)
-        return
-
-    def test_derandomize_wells(self) -> None:
-        A = (6, 8)
-        S = 13
-        randomizer = transform.WellRandomizer(A, S)
-        original = ["A01", "F02", "A05", "A07", "B07", "D06"]
-        expected = ["A01", "A02", "A03", "A04", "A05", "A06"]
-        derandomized = randomizer.derandomize_wells(original)
-        numpy.testing.assert_array_equal(expected, derandomized)
-        return
-
-    def test_derandomize_wells_bug_29(self) -> None:
-        A = (6, 8)
-        S = 13
-        randomizer = transform.WellRandomizer(A, S)
-        original = ["A01", "F02", "A05", "A07", "B07", "D06"][::-1]
-        expected = ["A01", "A02", "A03", "A04", "A05", "A06"][::-1]
-        derandomized = randomizer.derandomize_wells(original)
-        numpy.testing.assert_array_equal(expected, derandomized)
-        return
-
-    def test_randomize_wells_in_row(self) -> None:
-        A = (6, 8)
-        S = 13
-        randomizer = transform.WellRandomizer(A, S, mode="row")
-        original = ["A01", "A02", "A03", "B01", "C02", "B04"]
-        expected = ["A02", "A05", "A04", "B08", "C05", "B06"]
-        randomized = randomizer.randomize_wells(original)
-        numpy.testing.assert_array_equal(expected, randomized)
-        return
-
-    def test_derandomize_wells_in_row(self) -> None:
-        A = (6, 8)
-        S = 13
-        randomizer = transform.WellRandomizer(A, S, mode="row")
-        original = ["A02", "A05", "A04", "B08", "C05", "B06"]
-        expected = ["A01", "A02", "A03", "B01", "C02", "B04"]
-        randomized = randomizer.derandomize_wells(original)
-        numpy.testing.assert_array_equal(expected, randomized)
-        return
-
-    def test_randomize_wells_in_column(self) -> None:
-        A = (6, 8)
-        S = 13
-        randomizer = transform.WellRandomizer(A, S, mode="column")
-        original = ["A01", "A02", "A03", "B01", "C02", "B04"]
-        expected = ["B01", "D02", "B03", "D01", "A02", "E04"]
-        randomized = randomizer.randomize_wells(original)
-        numpy.testing.assert_array_equal(expected, randomized)
-        return
-
-    def test_derandomize_wells_in_column(self) -> None:
-        A = (6, 8)
-        S = 13
-        randomizer = transform.WellRandomizer(A, S, mode="column")
-        original = ["B01", "D02", "B03", "D01", "A02", "E04"]
-        expected = ["A01", "A02", "A03", "B01", "C02", "B04"]
-        randomized = randomizer.derandomize_wells(original)
-        numpy.testing.assert_array_equal(expected, randomized)
         return
